se agrego archivo enums para separar los enums de las definiciones
