+++ src/enums.ts
@@ -1,0 +1,11 @@
+export enum AirlineName {
+  EVA_Air = "EVA Air",
+  Linea_Aerea_Amaszonas = "Linea Aerea Amaszonas",
+  Thai_Airways = "Thai Airways",
+  Singapore_Airlines = "Singapore Airlines",
+  Air_Transat = "Air Transat",
+  TUI_Airways = "TUI Airways",
+  Star_Peru = "Star Peru",
+  Ukraine_International = "Ukraine International",
+  Turkish_Airlines = "Turkish Airlines"
+}
