+++ src/routes/diaries.ts
@@ -1,5 +1,6 @@
 import express from 'express';
 import * as diaryServices from '../services/diaryServices';
+import toNewDiaryEntry from '../utils';
 
 const router = express.Router();
 
@@ -30,18 +31,15 @@ router.get('/nside/:id', (req, res) => {
 })
 
 router.post('/', (req, res) => {
-  const { date, airline, flightNumber, seat } = req.body;
+  try {
+    const newDiaryEntry = toNewDiaryEntry(req.body);
 
-  const newDiaryEntry = diaryServices.addDiary(
-    {
-      date,
-      airline,
-      flightNumber,
-      seat
-    }
-  );
+    const addedDiaryEntry = diaryServices.addDiary(newDiaryEntry);
 
-  res.json(newDiaryEntry);
+    res.json(addedDiaryEntry);
+  } catch (e: any) {
+    res.status(400).send(e.message);
+  }
 })
 
 export default router;
